refactor(node): extract proof_history storage init into storage_init.rs
